+++ ninja-core-demo/src/main/java/conf/Module.java
@@ -16,26 +16,20 @@
 
 package conf;
 
-import javax.servlet.ServletContext;
-
-import ninja.NinjaAppAbstractModule;
-import ninja.NinjaServletDispatcher;
-
-import com.google.inject.servlet.ServletModule;
+import com.google.inject.AbstractModule;
 
 import etc.GreetingService;
 import etc.GreetingServiceImpl;
-import filters.DemoServletFilter;
 
-public class Module extends NinjaAppAbstractModule{
+public class Module extends AbstractModule {
 
-    public Module(ServletContext servletContext) {
-        super(servletContext);     
+    public Module() {
+        super();     
     }
 
   
     @Override
-    protected void setup() {       
+    protected void configure() {       
         // /////////////////////////////////////////////////////////////////////
         // Some guice bindings
         // /////////////////////////////////////////////////////////////////////
@@ -43,27 +37,7 @@ public class Module extends NinjaAppAbstractModule{
         bind(GreetingService.class).to(GreetingServiceImpl.class);
         // Bind the UDP ping controller so it starts up on server start
         // bind(UdpPingController.class);
-        
-    }
 
-    @Override
-    protected ServletModule setupServlets() {
-        // add new Servlet filter ONLY if you have some stuff that MUST use Servlet filter.
-        // otherwise use filters that provide ninja
-        
-        // every filter must be defined as singleton       
-        bind(NinjaServletDispatcher.class).asEagerSingleton();
-
-        //this one only as reference
-        //remove it in your app
-        bind(DemoServletFilter.class).asEagerSingleton();
-        return new ServletModule() {
-            @Override
-            protected void configureServlets() {                
-                filter("/*").through(DemoServletFilter.class);
-                serve("/*").with(NinjaServletDispatcher.class);
-            }
-        };
     }
 
 }