+++ ninja-core-test/src/main/java/ninja/NinjaRouterTest.java
@@ -17,6 +17,7 @@
 package ninja;
 
 import static org.junit.Assert.assertTrue;
+import ninja.servlet.NinjaBootstap;
 import ninja.utils.NinjaConstant;
 
 import org.junit.After;
@@ -40,7 +41,7 @@ public class NinjaRouterTest {
      */
     public void startServer() {
         // in this case servletContext can be null
-        NinjaBootstap ninjaBootup = new NinjaBootstap(null);
+        NinjaBootstap ninjaBootup = new NinjaBootstap();
         ninjaBootup.boot();
 
         router = ninjaBootup.getInjector().getInstance(Router.class);