+++ ninja-core/src/main/java/ninja/servlet/NinjaServletDispatcher.java
@@ -14,16 +14,21 @@
  * limitations under the License.
  */
 
-package ninja;
+package ninja.servlet;
 
 import java.io.IOException;
 
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
 import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import ninja.servlet.ContextImpl;
+import ninja.Context;
+import ninja.Ninja;
 
 import com.google.inject.Inject;
 import com.google.inject.Injector;
@@ -37,9 +42,7 @@ import com.google.inject.Injector;
  * @author ra
  * 
  */
-public class NinjaServletDispatcher extends HttpServlet {
-
-    private static final long serialVersionUID = 1L;
+public class NinjaServletDispatcher implements Filter {
 
     @Inject
     private Injector injector;
@@ -59,12 +62,19 @@ public class NinjaServletDispatcher extends HttpServlet {
         this.injector = injector;
     }
 
-   
-
-    
     @Override
-    public void service(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException{
+    public void init(FilterConfig filterConfig) throws ServletException {
+ 
+    }
+
+    @Override
+    public void doFilter(ServletRequest req,
+                         ServletResponse resp,
+                         FilterChain chain) throws IOException,
+            ServletException {
+
+        HttpServletRequest request = (HttpServletRequest) req;
+        HttpServletResponse response = (HttpServletResponse) resp;
 
         // We generate a Ninja compatible context element
         ContextImpl context = (ContextImpl) injector.getProvider(Context.class)
@@ -76,6 +86,11 @@ public class NinjaServletDispatcher extends HttpServlet {
         // And invoke ninja on it.
         // Ninja handles all defined routes, filters and much more:
         ninja.invoke(context);
+
+    }
+
+    @Override
+    public void destroy() {
 
     }