+++ ninja-core/src/main/java/ninja/validation/JSR303Validation.java
@@ -23,9 +23,9 @@ import java.lang.annotation.Target;
 
 /**
  * Validates that the annoted element is conform to its JSR303-Annotations
- * 
+ *
  * @author psommer
- * 
+ *
  */
 @WithValidator(Validators.JSRValidator.class)
 @Retention(RetentionPolicy.RUNTIME)
@@ -33,24 +33,17 @@ import java.lang.annotation.Target;
 public @interface JSR303Validation {
     /**
      * The key for the violation message
-     * 
+     *
      * @return The key of the violation message
      */
     String key() default KEY;
 
     /**
-     * Default message if the key isn't found
-     * 
+     * Default message if the field isn't found
+     *
      * @return The default message
      */
     String message() default MESSAGE;
-
-    /**
-     * The key for formatting the field
-     * 
-     * @return The key
-     */
-    String fieldKey() default "";
 
     public static final String KEY = "validation.is.JSR303.violation";
     public static final String MESSAGE = "{0} cannot be validated with JSR303 annotations";